fix parent menu item in menu edit form
